requirements.txt
anyio==3.6.2
appnope==0.1.3
asttokens==2.2.1
autopep8==2.0.2
backcall==0.2.0
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fastapi==0.95.1
Flask==2.2.5
gunicorn==20.1.0
idna==3.4
ipykernel==6.22.0
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pydantic==1.10.7
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.0.2
requests==2.30.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.26.1
tornado==6.3.1
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==2.0.2
wcwidth==0.2.6
Werkzeug==2.3.3